pyproject.toml
[tool.poetry]
name = "airbnb-clone"
version = "0.1.0"
description = ""
authors = ["ParkMinWoo-hub <136790431+ParkMinWoo-hub@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
